// frontend/orval/api/model/reservation.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { Room } from './room';

export interface Reservation {
  cost?: number;
  endTime?: string;
  id?: number;
  multi?: boolean;
  phoneNumber?: string;
  reserverName?: string;
  room?: Room;
  startTime?: string;
}